requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coloredlogs==15.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Cython==0.29.23
Django==3.2
djangorestframework==3.12.4
humanfriendly==9.1
idna==2.10
img2pdf==0.4.0
itypes==1.2.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
numpy==1.20.2
openapi-codec==1.3.2
opencv-python==4.5.1.48
pdfminer.six==20201018
pikepdf==2.11.4
Pillow==8.2.0
pkg-resources==0.0.0
pluggy==0.13.1
pycparser==2.20
PyPDF2==1.26.0
pytesseract==0.3.7
pytz==2021.1
regex==2021.4.4
reportlab==3.5.67
requests==2.25.1
simplejson==3.17.2
sortedcontainers==2.3.0
soupsieve==2.2.1
sqlparse==0.4.1
tqdm==4.60.0
uritemplate==3.0.1
urllib3==1.26.4
